requirements.txt
onnxruntime==1.9.0
torchvision==0.8.1
numpy==1.18.1
scipy==1.4.1
Pillow==7.1.2
streamlit==1.0.0


